Moved new trigger filters tests to rekt from experimental
